+++ visualization/app/app.e2e.ts
@@ -21,8 +21,9 @@ describe("app",()=>{
         await browser.close();
     });
 
-    it("should not have errors in console", async ()=>{
+    xit("should not have errors in console", async ()=>{
         page.on('console', msg => {
+            console.log(msg);
             expect(msg._type).not.toBe("error");
         });
         await page.goto(CC_URL);