+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeRendererService.ts
@@ -1,14 +1,20 @@
 import * as THREE from "three";
 import {WebGLRenderer} from "three";
+import {SettingsService, Settings, SettingsServiceSubscriber} from "../../../core/settings/settings.service";
 
 /**
  * A service which manages the Three.js renderer in an angular way.
  */
-export class ThreeRendererService {
+export class ThreeRendererService implements SettingsServiceSubscriber {
 
     public static SELECTOR = "threeRendererService";
 
-    public static CLEAR_COLOR = 0xeeeedd;
+    public static BACKGROUND_COLOR = {
+        white: 0xffffff,
+        normal: 0xeeeedd,
+    };
+
+    public static CLEAR_COLOR = ThreeRendererService.BACKGROUND_COLOR.normal;
 
     public static CLEAR_ALPHA = 1;
 
@@ -17,19 +23,36 @@ export class ThreeRendererService {
         preserveDrawingBuffer: true
     };
 
-
-    renderer: WebGLRenderer;
+    public renderer: WebGLRenderer;
 
     /* @ngInject */
-    constructor() {}
+    constructor(
+        private settingsService: SettingsService
+    ) {
+    }
 
     /**
      * Inits the renderer.
      */
     init(containerWidth: number, containerHeight: number){
         this.renderer = new THREE.WebGLRenderer(ThreeRendererService.RENDER_OPTIONS);
+        this.settingsService.subscribe(this);
+        this.onSettingsChanged(this.settingsService.settings, null);
+        this.setCurrentClearColorFromSettings(this.settingsService.settings);
         this.renderer.setSize(containerWidth, containerHeight);
         this.renderer.setClearColor(ThreeRendererService.CLEAR_COLOR, ThreeRendererService.CLEAR_ALPHA);
     }
 
+    public setCurrentClearColorFromSettings(settings: Settings) {
+        if (settings.isWhiteBackground) {
+            ThreeRendererService.CLEAR_COLOR = ThreeRendererService.BACKGROUND_COLOR.white;
+        } else {
+            ThreeRendererService.CLEAR_COLOR = ThreeRendererService.BACKGROUND_COLOR.normal;
+        }
+    }
+
+    onSettingsChanged(settings: Settings, event) {
+        this.setCurrentClearColorFromSettings(settings);
+        this.renderer.setClearColor(ThreeRendererService.CLEAR_COLOR, ThreeRendererService.CLEAR_ALPHA);
+    }
 }