+++ visualization/app/codeCharta/ui/colorSettingsPanel/colorSettingsPanel.component.ts
@@ -6,7 +6,8 @@ export class ColorSettingsPanelController implements SettingsServiceSubscriber{
     public viewModel = {
         flipped: false,
         deltas: false,
-        deltaColorFlipped: false
+        deltaColorFlipped: false,
+        backgroundFlipped: false
     };
 
     /* @ngInject */
@@ -21,11 +22,13 @@ export class ColorSettingsPanelController implements SettingsServiceSubscriber{
         this.viewModel.flipped = settings.neutralColorRange.flipped;
         this.viewModel.deltas = settings.deltas;
         this.viewModel.deltaColorFlipped = settings.deltaColorFlipped;
+        this.viewModel.backgroundFlipped = settings.isWhiteBackground;
     }
 
     apply() {
         this.settingsService.settings.neutralColorRange.flipped = this.viewModel.flipped;
         this.settingsService.settings.deltaColorFlipped = this.viewModel.deltaColorFlipped;
+        this.settingsService.settings.isWhiteBackground = this.viewModel.backgroundFlipped;
         this.settingsService.onSettingsChanged();
     }